Added face to character class, can enable/disable w/ enableFace
