+++ public/scripts/main.js
@@ -1,7 +1,7 @@
 import {Wall} from './Wall.js';
 import {Environment} from './Environment.js';
 import {Player, NPC} from './Character.js';
-import {DIRECTION, httpGet} from './helper_functions.js';
+import {DIRECTION} from './helper_functions.js';
 
 // Set premium content visbility
 const premiumContainer = document.getElementById('premium_content');
@@ -187,24 +187,10 @@ const npc = new NPC({
   colour: 'yellow',
   impassible: true,
   hp: 100,
+  enableFace: true,
 });
 npc.isSeeing(player);
 layer.add(npc.render);
-
-
-// Draw a face using cool-ascii-faces API
-const text = httpGet('https://fathomless-temple-39382.herokuapp.com/');
-const face = new Konva.Text({
-  x: 380,
-  y: 40,
-  text: (text.length < 14) ? text : '( ﾟヮﾟ)', // ensure we get a face that fits
-  fontSize: 16,
-  fill: '#555',
-  padding: 20,
-  align: 'center',
-});
-layer.add(face);
-
 
 const npcArray = [];
 npcArray.push(npc);
